symbol: mismatch size of variable lead to overflow in i386
store the negative value of the starting address of the module at offset
to calculate the relative address of the symbol later. unsigned long has
4byte size for i386. because of this, calculating between variable
offset and 8byte variable was not done as intended.

to explain this issue, see the belows:

utils/symbol.c:352
"offset -= iter.phdr.p_vaddr;"
(gdb) p/x iter.phdr.p_vaddr
$3 = 0x8048000
(gdb) p offset
$4 = 0
...
utils/symbol.c:229
(gdb) p/x elf_sym->st_value
$5 = 0x804a02c
(gdb) p/x offset
$6 = 0xf7fb8000
(gdb) ptype elf_sym->st_value
type = unsigned long long
(gdb) ptype offset
type = unsigned long
(gdb) p sizeof(offset)
$7 = 4
(gdb) p sizeof(elf_sym->st_value)
$8 = 8
(gdb) p sizeof(sym->addr)
$9 = 8
(gdb) ni
...
utils/symbol.c:230
sym->size = elf_sym->st_size;
(gdb) p/x sym->addr
$10 = 0x10000202c

as you can see, tried to substract the starting address of the module
stored in offset at elf_sym->st_value, but the offset and
elf_sym->st_value size were differred. so, compiler treat this operation
as 8-byte caculation. because of this, offset have treated as a positive
number instead non-negative number.

to solve this problem, increased size of offset to 8bytes.

+++ utils/symbol.c
@@ -201,7 +201,7 @@ static void unload_symtab(struct symtab *symtab)
 }
 
 static int load_symbol(struct symtab *symtab, unsigned long prev_sym_value,
-		       unsigned long offset, unsigned long flags,
+		       unsigned long long offset, unsigned long flags,
 		       struct uftrace_elf_data *elf,
 		       struct uftrace_elf_iter *iter)
 {
@@ -334,7 +334,7 @@ static void sort_symtab(struct symtab *symtab)
 }
 
 static int load_symtab(struct symtab *symtab, const char *filename,
-		       unsigned long offset, unsigned long flags)
+		       unsigned long long offset, unsigned long flags)
 {
 	int ret = -1;
 	unsigned long prev_sym_value = -1;
